feat: add validation in transaction operation

+++ src/main/java/com/example/bank/model/dto/TransactionCreateRequestDto.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class TransactionCreateRequestDto {
 	private Long accountId;
-	private int operationType;
+	private Long operationType;
 	private double amount;
 }
